+++ crates/mimic/src/design/mod.rs
@@ -13,9 +13,9 @@ pub mod prelude {
         core::{
             Key, Value,
             traits::{
-                EntityIdKind as _, EntityKind, FieldValue as _, Inner as _, Path as _,
-                Sanitize as _, Sanitizer as _, Serialize as _, Validate as _, ValidateCustom,
-                Validator as _, View as _, Visitable as _,
+                EntityKind, FieldValue as _, Inner as _, Path as _, Sanitize as _, Sanitizer as _,
+                Serialize as _, Validate as _, ValidateCustom, Validator as _, View as _,
+                Visitable as _,
             },
             view::View,
         },