+++ crates/mimic_declare/src/traits/schema_traits.rs
@@ -56,7 +56,6 @@ pub enum SchemaNodeKind {
     Newtype,
     Record,
     Sanitizer,
-    Selector,
     Set,
     Store,
     Tuple,