+++ crates/mimic_declare/src/view/field.rs
@@ -1,8 +1,0 @@
-impl HasViewExpr for FieldList {
-    fn view_type_expr(&self) -> TokenStream {
-        let fields = self.fields.iter().map(HasViewExpr::view_type_expr);
-        quote! {
-            #(#fields),*
-        }
-    }
-}